Move all of the Seed functionality into its own file
