+++ random.cabal
@@ -85,6 +85,7 @@ library
     exposed-modules:
         System.Random
         System.Random.Internal
+        System.Random.Seed
         System.Random.Stateful
     other-modules:
         System.Random.GFinite
