+++ src/System/Random.hs
@@ -30,12 +30,6 @@ module System.Random
       , genWord64R
       , unsafeUniformFillMutableByteArray
       )
-  , SeedGen (..)
-  , Seed
-  , mkSeed
-  , mkSeedFromByteString
-  , unSeed
-  , seedSize
   , SplitGen (splitGen)
   , uniform
   , uniformR
@@ -43,6 +37,8 @@ module System.Random
   , Uniform
   , UniformRange
   , Finite
+  -- ** Seed
+  , module System.Random.Seed
   -- * Generators for sequences of pseudo-random bytes
   -- ** Lists
   , uniforms
@@ -100,6 +96,7 @@ import Foreign.C.Types
 import GHC.Exts
 import System.Random.GFinite (Finite)
 import System.Random.Internal
+import System.Random.Seed
 import qualified System.Random.SplitMix as SM
 
 -- $introduction
