+++ src/System/Random/Internal.hs
@@ -1,24 +1,19 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE CPP #-}
-{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DefaultSignatures #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GHCForeignImportPrim #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MagicHash #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE Trustworthy #-}
-{-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeFamilyDependencies #-}
 {-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE TypeFamilyDependencies #-}
 {-# LANGUAGE UnboxedTuples #-}
 {-# LANGUAGE UndecidableInstances #-}
-{-# LANGUAGE UndecidableSuperClasses #-}
 {-# LANGUAGE UnliftedFFITypes #-}
 {-# OPTIONS_HADDOCK hide, not-home #-}
 
@@ -34,17 +29,7 @@ module System.Random.Internal
   (-- * Pure and monadic pseudo-random number generator interfaces
     RandomGen(..)
   , SplitGen(..)
-  , SeedGen(..)
-  -- ** Seed
   , Seed(..)
-  , seedSize
-  , mkSeed
-  , unSeed
-  , mkSeedFromByteString
-  , unSeedToByteString
-  , withSeedFile
-  , nonEmptyToSeed
-  , nonEmptyFromSeed
   -- * Stateful
   , StatefulGen(..)
   , FrozenGen(..)
@@ -94,33 +79,37 @@ module System.Random.Internal
   , genByteArrayST
   , genShortByteStringIO
   , genShortByteStringST
+  , defaultUnsafeFillMutableByteArrayT
   , defaultUnsafeUniformFillMutableByteArray
   -- ** Helpers for dealing with MutableByteArray
   , newMutableByteArray
   , newPinnedMutableByteArray
   , freezeMutableByteArray
   , writeWord8
+  , writeWord64LE
+  , indexWord8
+  , indexWord64LE
+  , indexByteSliceWord64LE
+  , sizeOfByteArray
+  , shortByteStringToByteArray
+  , byteArrayToShortByteString
   ) where
 
 import Control.Arrow
 import Control.DeepSeq (NFData)
-import Control.Monad (replicateM, when, unless, (>=>))
+import Control.Monad (replicateM, when, (>=>))
 import Control.Monad.Cont (ContT, runContT)
 import Control.Monad.Identity (IdentityT (runIdentityT))
 import Control.Monad.ST
-import Control.Monad.IO.Class
 import Control.Monad.State.Strict (MonadState(..), State, StateT(..), execStateT, runState)
 import Control.Monad.Trans (lift, MonadTrans)
 import Data.Array.Byte (ByteArray(..), MutableByteArray(..))
 import Data.Bits
-import qualified Data.ByteString as BS
 import Data.ByteString.Short.Internal (ShortByteString(SBS))
-import qualified Data.ByteString.Short as SBS (fromShort, toShort)
+import qualified Data.ByteString.Short as SBS (fromShort)
 import Data.IORef (IORef, newIORef)
 import Data.Int
 import Data.List (sortOn)
-import Data.List.NonEmpty (NonEmpty(..), nonEmpty)
-import Data.Typeable
 import Data.Word
 import Foreign.C.Types
 import Foreign.Storable (Storable)
@@ -128,7 +117,6 @@ import GHC.Exts
 import GHC.Generics
 import GHC.IO (IO(..))
 import GHC.ST (ST(..))
-import GHC.TypeLits (Nat, KnownNat, natVal, type (<=))
 import GHC.Word
 import Numeric.Natural (Natural)
 import System.IO.Unsafe (unsafePerformIO)
@@ -145,6 +133,19 @@ import Data.ByteString (ByteString)
 
 -- Needed for WORDS_BIGENDIAN
 #include "MachDeps.h"
+
+-- | This is a binary form of pseudo-random number generator's state. It is designed to be
+-- safe and easy to use for input/output operations like restoring from file, transmitting
+-- over the network, etc.
+--
+-- Constructor is not exported, becasue it is important for implementation to enforce the
+-- invariant of the unrelying byte array being of exact same length as the generator has
+-- specified in `System.Random.Seed.SeedSize`. Used `System.Random.Seed.mkSize` and
+-- `System.Random.Seed.unSize` to get access to the raw bytes in a safe manner.
+--
+-- @since 1.3.0
+newtype Seed g = Seed ByteArray
+  deriving (Eq, Ord, Show)
 
 
 -- | 'RandomGen' is an interface to pure pseudo-random number generators.
@@ -299,138 +300,6 @@ class RandomGen g => SplitGen g where
   -- @since 1.3.0
   splitGen :: g -> (g, g)
 
--- | This form of the pseudo-random is designed to be to restore from and easy to serialize.
---
--- @since 1.3.0
-newtype Seed g = Seed ByteArray
-  deriving (Eq, Ord)
-
--- | Get the expected size in bytes of the `Seed`
---
--- @since 1.3.0
-seedSize :: forall g. SeedGen g => Int
-seedSize = fromIntegral $ natVal (Proxy :: Proxy (SeedSize g))
-
--- | Construct a `Seed` from a `ByteArray` of expected length. Whenever `ByteArray` does
--- not match the `SeedSize` specified by the pseudo-random generator, this function will
--- return `Nothing`
---
--- @since 1.3.0
-mkSeed :: forall g m. (SeedGen g, MonadFail m) => ByteArray -> m (Seed g)
-mkSeed ba = do
-  unless (sizeOfByteArray ba == seedSize @g) $ do
-    fail $ "Unexpected number of bytes: "
-        <> show (sizeOfByteArray ba)
-        <> ". Exactly "
-        <> show (seedSize @g)
-        <> " bytes is required by the "
-        <> show (genTypeName @g)
-  pure $ Seed ba
-
--- | Just like `mkSeed`, but uses `ByteString` as argument. Results in a memcopy of the seed.
---
--- @since 1.3.0
-mkSeedFromByteString :: (SeedGen g, MonadFail m) => ByteString -> m (Seed g)
-mkSeedFromByteString = mkSeed . shortByteStringToByteArray . SBS.toShort
-
--- | Unwrap the `Seed` and get the underlying `ByteArray`
---
--- @since 1.3.0
-unSeed :: Seed g -> ByteArray
-unSeed (Seed ba) = ba
-
--- | Just like `unSeed`, but produced a `ByteString`. Results in a memcopy of the seed.
---
--- @since 1.3.0
-unSeedToByteString :: Seed g -> ByteString
-unSeedToByteString = SBS.fromShort . byteArrayToShortByteString . unSeed
-
-
--- | Read the seed from a file and use it for constructing a pseudo-random number
--- generator. After supplied action has been applied to the constructed generator, the
--- resulting generator will be converted back to a seed and written to the same file.
---
--- @since 1.3.0
-withSeedFile :: (SeedGen g, MonadIO m) => FilePath -> (g -> m (a, g)) -> m a
-withSeedFile fileName f = do
-  bs <- liftIO $ BS.readFile fileName
-  seed <- liftIO $ mkSeedFromByteString bs
-  (res, gen) <- f $ seedGen seed
-  liftIO $ BS.writeFile fileName $ unSeedToByteString $ unseedGen gen
-  pure res
-
-
-nonEmptyToSeed :: forall g. SeedGen g => NonEmpty Word64 -> Seed g
-nonEmptyToSeed xs = Seed $ runST $ do
-  let n = seedSize @g
-  mba <- newMutableByteArray n
-  _ <- flip runStateT (toList xs) $ do
-    defaultUnsafeFillMutableByteArrayT mba 0 n $ do
-      get >>= \case
-        [] -> pure 0
-        w:ws -> w <$ put ws
-  freezeMutableByteArray mba
-
-nonEmptyFromSeed :: forall g. SeedGen g => Seed g -> NonEmpty Word64
-nonEmptyFromSeed (Seed ba) =
-  case nonEmpty $ reverse $ goWord64 0 [] of
-    Just ne -> ne
-    Nothing -> -- Seed is at least 1 byte in size, so it can't be empty
-      error $ "Impossible: Seed must be at least: "
-           ++ show (seedSize @g)
-           ++ " bytes, but got "
-           ++ show n
-  where
-    n = sizeOfByteArray ba
-    n8 = 8 * (n `quot` 8)
-    goWord64 i !acc
-      | i < n8 = goWord64 (i + 8) (indexWord64LE ba i : acc)
-      | i == n = acc
-      | otherwise = indexByteSliceWord64LE ba i n : acc
-
--- | Interface for coverting a pure pseudo-random number generator to and from non-empty
--- sequence of bytes/words. Seeds are stored in Little-Endian order regardless of the platform
--- it is being used on, which provides inter-platform compatibility, while providing
--- optimal performance for most common platforms.
---
--- Conversion to and from a `Seed` serves as a building block for implementing
--- serialization for any pure or frozen pseudo-random number generator
---
--- @since 1.3.0
-class (KnownNat (SeedSize g), 1 <= SeedSize g) => SeedGen g where
-  type SeedSize g :: Nat
-  {-# MINIMAL (seedGen, unseedGen)|(seedGen64, unseedGen64) #-}
-
-  -- | Convert from a binary representation to a pseudo-random number generator
-  --
-  -- @since 1.3.0
-  seedGen :: Seed g -> g
-  seedGen = seedGen64 . nonEmptyFromSeed
-
-  -- |
-  --
-  -- @since 1.3.0
-  unseedGen :: g -> Seed g
-  unseedGen = nonEmptyToSeed . unseedGen64
-
-  -- |
-  --
-  -- @since 1.3.0
-  seedGen64 :: NonEmpty Word64 -> g
-  seedGen64 = seedGen . nonEmptyToSeed
-
-  -- |
-  --
-  -- @since 1.3.0
-  unseedGen64 :: g -> NonEmpty Word64
-  unseedGen64 = nonEmptyFromSeed . unseedGen
-
-  -- | This is a function that shows the name of the generator type, which is useful for
-  -- error reporting. If you need to avoid the `Typeable` constraint then feel free to
-  -- override the default definition.
-  genTypeName :: String
-  default genTypeName :: Typeable g => String
-  genTypeName = show (typeOf (Proxy @g))
 
 
 -- | 'StatefulGen' is an interface to monadic pseudo-random number generators.
@@ -942,7 +811,7 @@ data StateGenM g = StateGenM
 --
 -- @since 1.2.0
 newtype StateGen g = StateGen { unStateGen :: g }
-  deriving (Eq, Ord, Show, SeedGen, RandomGen, Storable, NFData)
+  deriving (Eq, Ord, Show, RandomGen, Storable, NFData)
 
 instance (RandomGen g, MonadState g m) => StatefulGen (StateGenM g) m where
   uniformWord32R r _ = state (genWord32R r)
@@ -1098,23 +967,10 @@ shuffleListM xs gen = do
 
 -- | The standard pseudo-random number generator.
 newtype StdGen = StdGen { unStdGen :: SM.SMGen }
-  deriving (Show, SeedGen, RandomGen, SplitGen, NFData)
+  deriving (Show, RandomGen, SplitGen, NFData)
 
 instance Eq StdGen where
   StdGen x1 == StdGen x2 = SM.unseedSMGen x1 == SM.unseedSMGen x2
-
-
-instance SeedGen SM.SMGen where
-  type SeedSize SM.SMGen = 16
-  seedGen (Seed ba) =
-    SM.seedSMGen (indexWord64LE ba 0) (indexWord64LE ba 8)
-  unseedGen g =
-    case SM.unseedSMGen g of
-      (seed, gamma) -> Seed $ runST $ do
-        mba <- newMutableByteArray 16
-        writeWord64LE mba 0 seed
-        writeWord64LE mba 8 gamma
-        freezeMutableByteArray mba
 
 instance RandomGen SM.SMGen where
   next = SM.nextInt
@@ -1132,24 +988,6 @@ instance RandomGen SM.SMGen where
 instance SplitGen SM.SMGen where
   splitGen = SM.splitSMGen
   {-# INLINE splitGen #-}
-
-instance SeedGen SM32.SMGen where
-  type SeedSize SM32.SMGen = 8
-  seedGen (Seed ba) =
-    let x = indexWord64LE ba 0
-        seed, gamma :: Word32
-        seed = fromIntegral (shiftR x 32)
-        gamma = fromIntegral x
-    in SM32.seedSMGen seed gamma
-  unseedGen g =
-    let seed, gamma :: Word32
-        (seed, gamma) = SM32.unseedSMGen g
-    in Seed $ runST $ do
-        mba <- newMutableByteArray 8
-        let w64 :: Word64
-            w64 = shiftL (fromIntegral seed) 32 .|. fromIntegral gamma
-        writeWord64LE mba 0 w64
-        freezeMutableByteArray mba
 
 instance RandomGen SM32.SMGen where
   next = SM32.nextInt
